Modify :hidden variable constraint fn signature
The constraint function's arity would depend on the given list of
variables.
This is a bit restrictive because the constraint may be
generated dynamically, and along with the list, a matching fn needs to
be constructed as well.

Instead the fn could receive a list (or map) of assignments of the
variables. The author can choose to pattern match.

SA solver
---------

The SA solver perturbated all variables (including hidden ones). It is
impossible to find the neighbour of compound variables (unless the
constraint is very well understood).
Instead, the perturbation should be restricted only to the _real
variables_.